Optimization is not working

+++ src/stamina/StaminaModelBuilder.h
@@ -263,7 +263,7 @@ namespace stamina {
 		bool firstIteration;
 		double localKappa;
 		bool isCtmc;
-		uint32_t numberTerminal;
+		uint64_t numberTerminal;
 	};
